requirements.txt
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
blis==0.7.11
catalogue==1.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
docopt==0.6.2
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.0
huggingface-hub==0.29.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
plac==1.1.3
preshed==3.0.9
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sortedcontainers==2.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
sympy==1.13.1
thinc==7.4.6
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==0.10.1
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
